logging/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.92.0
h11==0.14.0
hazelcast-python-client==5.2.0
idna==3.4
pydantic==1.10.5
sniffio==1.3.0
starlette==0.25.0
typing-extensions==4.5.0
uvicorn==0.20.0
